+++ tests/Middleware/ForceHttpsMiddlewareTest.php
@@ -10,9 +10,9 @@ use AvtoDev\RoadRunnerLaravel\Tests\AbstractTestCase;
 use AvtoDev\RoadRunnerLaravel\Middleware\ForceHttpsMiddleware;
 
 /**
- * @group middleware
+ * @group  middleware
  *
- * @covers \AvtoDev\RoadRunnerLaravel\Middleware\ForceHttpsMiddleware
+ * @covers \AvtoDev\RoadRunnerLaravel\Middleware\ForceHttpsMiddleware<extended>
  */
 class ForceHttpsMiddlewareTest extends AbstractTestCase
 {
@@ -40,7 +40,7 @@ class ForceHttpsMiddlewareTest extends AbstractTestCase
     /**
      * @return void
      */
-    public function testForceHttpsSchemaIfHeaderArePresents()
+    public function testForceHttpsSchemaIfHeaderArePresents(): void
     {
         ($request = new Request)->headers->set('HTTPS', 'HTTPS');
 
@@ -64,7 +64,7 @@ class ForceHttpsMiddlewareTest extends AbstractTestCase
     /**
      * @return void
      */
-    public function testHttpsSchemaNotForcedIfHeaderAreNotPresents()
+    public function testHttpsSchemaNotForcedIfHeaderAreNotPresents(): void
     {
         $request = new Request;