+++ tests/Support/Stacks/CallbacksStackTest.php
@@ -7,12 +7,12 @@ namespace AvtoDev\RoadRunnerLaravel\Tests\Support\Stacks;
 use AvtoDev\RoadRunnerLaravel\Support\Stacks\CallbacksStack;
 
 /**
- * @group stacks
+ * @group  stacks
  *
- * @covers \AvtoDev\RoadRunnerLaravel\Support\Stacks\CallbacksStack
- * @covers \AvtoDev\RoadRunnerLaravel\Support\Stacks\AbstractStack
+ * @covers \AvtoDev\RoadRunnerLaravel\Support\Stacks\CallbacksStack<extended>
+ * @covers \AvtoDev\RoadRunnerLaravel\Support\Stacks\AbstractStack<extended>
  */
-class CallbacksStackTest extends BasicStackTest
+class CallbacksStackTest extends BasicStackTestCase
 {
     /**
      * @var CallbacksStack
@@ -24,7 +24,7 @@ class CallbacksStackTest extends BasicStackTest
      *
      * @return void
      */
-    public function testMakeMethod()
+    public function testMakeMethod(): void
     {
         $callable = function () {
         };
@@ -39,7 +39,7 @@ class CallbacksStackTest extends BasicStackTest
      *
      * @return void
      */
-    public function testStackAccessors()
+    public function testStackAccessors(): void
     {
         $this->assertCount(0, $this->instance);
 
@@ -69,7 +69,7 @@ class CallbacksStackTest extends BasicStackTest
     /**
      * @return void
      */
-    public function testConstructorWithInvalidElements()
+    public function testConstructorWithInvalidElements(): void
     {
         $instance = $this->instanceFactory([
             'foo',
@@ -87,7 +87,7 @@ class CallbacksStackTest extends BasicStackTest
     /**
      * @return void
      */
-    public function testCallEach()
+    public function testCallEach(): void
     {
         $counter = 0;